+++ packages/x-components/src/x-modules/search/store/__tests__/actions.spec.ts
@@ -613,8 +613,8 @@ describe('testing search module actions', () => {
   });
 
   describe('setUrlParams', () => {
-    it('should set the params of the search module', async () => {
-      resetSearchStateWith(store, { query: 'funko', page: 1, sort: '' });
+    it('should set the params of the search module when the query is the same', async () => {
+      resetSearchStateWith(store, { query: 'lego', page: 1, sort: '' });
 
       await store.dispatch('setUrlParams', {
         query: 'lego',
@@ -627,22 +627,37 @@ describe('testing search module actions', () => {
       expect(store.state.sort).toEqual('priceSort asc');
     });
 
+    // eslint-disable-next-line max-len
+    it('should set the params of the search module, except page, when the query is different', async () => {
+      resetSearchStateWith(store, { query: 'funko', page: 1, sort: '' });
+
+      await store.dispatch('setUrlParams', {
+        query: 'lego',
+        page: 2,
+        sort: 'priceSort asc'
+      } as UrlParams);
+
+      expect(store.state.query).toEqual('lego');
+      expect(store.state.page).toEqual(1);
+      expect(store.state.sort).toEqual('priceSort asc');
+    });
+
     it('should set in the search module the query value even if empty', async () => {
       resetSearchStateWith(store, { query: 'funko' });
 
-      await store.dispatch('setUrlParams', { page: 2, query: '' } as UrlParams);
+      await store.dispatch('setUrlParams', { sort: 'priceSort asc', query: '' } as UrlParams);
 
       expect(store.state.query).toEqual('');
-      expect(store.state.page).toEqual(2);
+      expect(store.state.sort).toEqual('priceSort asc');
     });
 
     it('should set in the search module the sort value even if empty', async () => {
       resetSearchStateWith(store, { sort: 'priceSort asc' });
 
-      await store.dispatch('setUrlParams', { page: 2, sort: '' } as UrlParams);
+      await store.dispatch('setUrlParams', { page: 1, sort: '' } as UrlParams);
 
       expect(store.state.sort).toEqual('');
-      expect(store.state.page).toEqual(2);
+      expect(store.state.page).toEqual(1);
     });
   });